Safety Improvements
QUERY_STRING variables are no longer trusted implicitly. Every variable
that is read from the query string is prefixed with qs_. Only variables
which the script is actually expecting to use are then cast to their
proper name without the qs_ prefix.

A script should otherwise never use the qs_ version of a variable.